make all the *ApiClients interfaces for consistency and for removing the @Mockable thing. Clean up OAuthApiClient API.
