+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osm/edits/upload/changesets/ChangesetApiClient.kt
@@ -3,25 +3,8 @@ package de.westnordost.streetcomplete.data.osm.edits.upload.changesets
 import de.westnordost.streetcomplete.data.AuthorizationException
 import de.westnordost.streetcomplete.data.ConflictException
 import de.westnordost.streetcomplete.data.ConnectionException
-import de.westnordost.streetcomplete.data.user.UserAccessTokenSource
-import de.westnordost.streetcomplete.data.wrapApiClientExceptions
-import de.westnordost.streetcomplete.util.Mockable
-import io.ktor.client.HttpClient
-import io.ktor.client.call.body
-import io.ktor.client.plugins.ClientRequestException
-import io.ktor.client.plugins.expectSuccess
-import io.ktor.client.request.bearerAuth
-import io.ktor.client.request.put
-import io.ktor.client.request.setBody
-import io.ktor.http.HttpStatusCode
 
-@Mockable
-class ChangesetApiClient(
-    private val httpClient: HttpClient,
-    private val baseUrl: String,
-    private val userAccessTokenSource: UserAccessTokenSource,
-    private val serializer: ChangesetApiSerializer,
-) {
+interface ChangesetApiClient {
     /**
      * Open a new changeset with the given tags
      *
@@ -33,14 +16,7 @@ class ChangesetApiClient(
      *
      * @return the id of the changeset
      */
-    suspend fun open(tags: Map<String, String>): Long = wrapApiClientExceptions {
-        val response = httpClient.put(baseUrl + "changeset/create") {
-            userAccessTokenSource.accessToken?.let { bearerAuth(it) }
-            setBody(serializer.serialize(tags))
-            expectSuccess = true
-        }
-        return response.body<String>().toLong()
-    }
+    suspend fun open(tags: Map<String, String>): Long
 
     /**
      * Closes the given changeset.
@@ -52,19 +28,6 @@ class ChangesetApiClient(
      *                                (OAuth scope "write_api")
      * @throws ConnectionException if a temporary network connection problem occurs
      */
-    suspend fun close(id: Long): Unit = wrapApiClientExceptions {
-        try {
-            httpClient.put(baseUrl + "changeset/$id/close") {
-                userAccessTokenSource.accessToken?.let { bearerAuth(it) }
-                expectSuccess = true
-            }
-        } catch (e: ClientRequestException) {
-            when (e.response.status) {
-                HttpStatusCode.Conflict, HttpStatusCode.NotFound -> {
-                    throw ConflictException(e.message, e)
-                }
-                else -> throw e
-            }
-        }
-    }
+    suspend fun close(id: Long)
 }
+
