+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/osmcal/CalendarEventsModule.kt
@@ -3,7 +3,7 @@ package de.westnordost.streetcomplete.data.osmcal
 import org.koin.dsl.module
 
 val calendarEventsModule = module {
-    factory { OsmCalApiClient(get(), get()) }
+    factory<OsmCalApiClient> { OsmCalApiClientImpl(get(), get()) }
     factory { OsmCalParser() }
     factory { OsmCalUpdater(get(), get(), get()) }
     factory { CalendarEventsDao(get()) }
