+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/user/UserApiClient.kt
@@ -2,42 +2,18 @@ package de.westnordost.streetcomplete.data.user
 
 import de.westnordost.streetcomplete.data.AuthorizationException
 import de.westnordost.streetcomplete.data.ConnectionException
-import de.westnordost.streetcomplete.data.wrapApiClientExceptions
-import de.westnordost.streetcomplete.util.Mockable
-import io.ktor.client.HttpClient
-import io.ktor.client.plugins.ClientRequestException
-import io.ktor.client.plugins.expectSuccess
-import io.ktor.client.request.bearerAuth
-import io.ktor.client.request.get
-import io.ktor.client.statement.bodyAsChannel
-import io.ktor.http.HttpStatusCode
-import io.ktor.utils.io.asSource
-import kotlinx.io.buffered
 
 /**
  * Talks with OSM user API
  */
-@Mockable
-class UserApiClient(
-    private val httpClient: HttpClient,
-    private val baseUrl: String,
-    private val userAccessTokenSource: UserAccessTokenSource,
-    private val userApiParser: UserApiParser,
-) {
+interface UserApiClient {
     /**
      * @return the user info of the current user
      *
      * @throws AuthorizationException if we are not authorized to read user details (scope "read_prefs")
      * @throws ConnectionException on connection or server error
      */
-    suspend fun getMine(): UserInfo = wrapApiClientExceptions {
-        val response = httpClient.get(baseUrl + "user/details") {
-            userAccessTokenSource.accessToken?.let { bearerAuth(it) }
-            expectSuccess = true
-        }
-        val source = response.bodyAsChannel().asSource().buffered()
-        return userApiParser.parseUsers(source).first()
-    }
+    suspend fun getMine(): UserInfo
 
     /**
      * @param userId id of the user to get the user info for
@@ -45,16 +21,6 @@ class UserApiClient(
      *
      * @throws ConnectionException on connection or server error
      */
-    suspend fun get(userId: Long): UserInfo? = wrapApiClientExceptions {
-        try {
-            val response = httpClient.get(baseUrl + "user/$userId") { expectSuccess = true }
-            val source = response.bodyAsChannel().asSource().buffered()
-            return userApiParser.parseUsers(source).first()
-        } catch (e: ClientRequestException) {
-            when (e.response.status) {
-                HttpStatusCode.Gone, HttpStatusCode.NotFound -> return null
-                else -> throw e
-            }
-        }
-    }
+    suspend fun get(userId: Long): UserInfo?
 }
+
