+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/weeklyosm/WeeklyOsmApiClientImpl.kt
@@ -1,0 +1,19 @@
+package de.westnordost.streetcomplete.data.weeklyosm
+
+import io.ktor.client.HttpClient
+import io.ktor.client.request.get
+import io.ktor.client.statement.bodyAsChannel
+import io.ktor.utils.io.asSource
+import kotlinx.io.buffered
+import kotlin.time.Instant
+
+class WeeklyOsmApiClientImpl(
+    private val httpClient: HttpClient,
+    private val parser: WeeklyOsmRssFeedParser,
+) : WeeklyOsmApiClient {
+    override suspend fun getLastPublishDate(): Instant? {
+        val response = httpClient.get("https://weeklyosm.eu/feed")
+        val source = response.bodyAsChannel().asSource().buffered()
+        return parser.parseLastPublishDate(source)
+    }
+}
