+++ app/src/commonMain/kotlin/de/westnordost/streetcomplete/data/weeklyosm/WeeklyOsmModule.kt
@@ -3,7 +3,7 @@ package de.westnordost.streetcomplete.data.weeklyosm
 import org.koin.dsl.module
 
 val weeklyOsmModule = module {
-    factory { WeeklyOsmApiClient(get(), get()) }
+    factory<WeeklyOsmApiClient> { WeeklyOsmApiClientImpl(get(), get()) }
     factory { WeeklyOsmUpdater(get(), get()) }
     factory { WeeklyOsmRssFeedParser() }
 }
