+++ app/src/commonTest/kotlin/de/westnordost/streetcomplete/data/osm/edits/upload/changesets/ChangesetApiClientImplTest.kt
@@ -12,7 +12,7 @@ import kotlin.test.assertFailsWith
 // other than some other APIs we are speaking to, we do not control the OSM API, so I think it is
 // more effective to test with the official test API instead of mocking some imagined server
 // response
-class ChangesetApiClientTest {
+class ChangesetApiClientImplTest {
 
     @Test fun `open throws exception on insufficient privileges`(): Unit = runBlocking {
         assertFailsWith<AuthorizationException> {
@@ -38,7 +38,7 @@ class ChangesetApiClientTest {
     }
 
     private fun client(token: String?) =
-        ChangesetApiClient(
+        ChangesetApiClientImpl(
             httpClient = HttpClient(),
             baseUrl = OsmDevApi.URL,
             userAccessTokenSource = object : UserAccessTokenSource { override val accessToken = token.orEmpty() },
